pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "final-sc-review"
version = "0.1.0"
description = "Final S–C evidence sentence retrieval pipeline with BGE-M3 and Jina-v3"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "CVR Lab"}]

dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "pyyaml>=6.0",
  "torch>=2.0",
  "transformers>=4.40",
  "tqdm>=4.66",
  "FlagEmbedding>=1.2.10",
  "optuna>=3.6",
  "pytest>=7.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]
norecursedirs = ["_archived"]
